Split out ArgsToFrontenfOptionsConverter.
